chore: move `demo.rs` to existing `test/sample.rs`
